moved the components into a directory and made a separate component to reuse the button
