--- src/app.typedefs.ts
@@ -1,10 +1,0 @@
-export enum SortType {
-  NONE,
-  ALPHABET,
-  LENGTH,
-}
-
-export interface ReorderOptions {
-  sortType: SortType,
-  isReversed: boolean,
-}
