+++ src/components/App/app.typedefs.ts
@@ -1,0 +1,10 @@
+export enum SortType {
+  NONE,
+  ALPHABET,
+  LENGTH,
+}
+
+export interface ReorderOptions {
+  sortType: SortType,
+  isReversed: boolean,
+}
